// src/app/interfaces/common/motto.interface.ts
export interface Moto {
  _id?: string;
  name?: string;
  nameEn?: string;
  image?: string;
  seoImage?: string;
  informations?: any[];
  description?: string;
  title?: string;
  descriptionEn?: string;
  titleEn?: string;
  subTitle?: string;
  shortDescription?: string;
  subTitleEn?: string;
  shortDescriptionEn?: string;
  slug?: string;
  createdAt?: Date;
  updatedAt?: Date;
}